daemon: finish threads on close
Current daemon shutdown can cause crashes. The problem is that threads
serving client request are joined on daemon dispose after drivers already
cleaned up. But this threads typically uses drivers and thus crashes come.
We need to join threads before virStateCleanup. virNetDaemonClose is
a good candidate.

But it turns out that we can hang on join. The problem is that at this
moment event loop is not functional and for example threads waiting for
qemu response will never finish. Let's introduce extra shutdown step
for drivers so that they can signal any API calls in progress to finish.
Message-Id: <1506516338-372467-2-git-send-email-nshirokovskiy@virtuozzo.com>