qemu: monitor: check monitor not closed on send
It is convinient to have closed state for qemu monitor. I'm going
to use it on daemon shutdown. We can use fd to check for this state.

Now we don't need to set lastError on qemuMonitorClose as the error will
be reported explicitly by qemuMonitorSend.

Let's report VIR_ERR_OPERATION_INVALID instead of VIR_ERR_OPERATION_FAILED
as closing monitor is a part of stopping domain (and will be a part
of shutting driver down). Particularly we won't see error messages
in daemon log.
Message-Id: <1506516338-372467-3-git-send-email-nshirokovskiy@virtuozzo.com>