docs, rng: Allow a pool name to be line domain name
https://bugzilla.redhat.com/show_bug.cgi?id=1475250

It's possible to define and start a pool with a '.' in the
name; however, when trying to add a volume to a domain using
the storage pool source with a name with a '.' in the name,
the domain RNG validation fails because RNG uses 'genericName'
which does not allow a '.' in the name. Pool definition has
no similar call to virXMLValidateAgainstSchema. Pool name
validation occurs in storagePoolDefineXML and only calls
virXMLCheckIllegalChars using the same parameter "\n" as
qemuDomainDefineXMLFlags would check after the RNG check
could be succesful.

So in order to resolve this, create a poolName definition
in the RNG and allow the pool name and the volume source
pool name to use that definition.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170927190735.18078-1-jferlan@redhat.com>