virDomainNetFindIdx: Ignore auto generated MAC addresses
When detaching an <interface/> from domain, it's MAC address is
parsed and if not present one is generated. If, however, no
corresponding interface is found in the domain, the following
error is reported:

error: operation failed: no device matching mac address 52:54:00:75:32:5b found

where the MAC address is the auto generated one. This might be
very confusing. Solution to this is to ignore auto generated MAC
address when looking up the device.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <d57e4baddd80c809cdf19959f755e4c137b3b7bb.1506941768.git.mprivozn@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com> to this patch.