ansible: Add initial support
This commit adds just enough to bootstrap Jenkins workers and install
packages required to build any project, such as gcc and make.

Ubuntu 12-16 are supported, even though they're not actually used in
libvirt CI, because there's a lot of overlap between this and Travis
CI, which only supports Ubuntu instead.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20171002151056.30841-2-abologna@redhat.com>