cpu: Return model from virCPUModelIsAllowed
If a given CPU model is supported by the hypervisor, we want to know
more about it, e.g., what features may block its usage on the current
host and such details are stored in the virDomainCapsCPUModelsPtr list
which virCPUModelIsAllowed uses to check whether the CPU model is
supported. Thus if the CPU model is found in the list we can directly
return a pointer to the corresponding virDomainCapsCPUModel if the
caller needs to look at the details.

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>
Message-Id: <a578d0ea532d6f501f491468962d809f16d845f7.1507128897.git.jdenemar@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>