cputest: Print correct feature in virCPUUpdateLive test
If the actual result does not match our expectation, the tests would
not correctly show the difference if a CPU feature is disabled in the
expected result and the actual result does not mention it at all. The
test could complain about an unrelated CPU feature or it could even
crash in case the actual result contains no more features to go through.

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>
Message-Id: <3ebf561d2d82308c93feb36fc8a4537dee8c5e5b.1507128897.git.jdenemar@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>