storage: Alter storageBackendCreateQemuImgSecretObject args
Since all that was really needed was a couple of fields and building
the object can be more generic, let's alter the args a bit. This will
be useful shortly for adding the secret object for a volume resize
operation on a luks volume that will need a secret object.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171006181332.16531-4-jferlan@redhat.com>