util: Introduce virFileWaitForExists
Since we have a number of places where we workaround timing issues with
devices, attributes (files in general) not being available at the time
of processing them by calling usleep in a loop for a fixed number of
tries, we could as well have a utility function that would do that.
Therefore we won't have to duplicate this ugly workaround even more.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <5f6095cf15006d1baa86bf6eb51b0d878ee127b4.1507732721.git.eskultet@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>