util: storage: Store backing chain index in virStorageSource
The backing store indexes were not bound to the storage sources in any
way. To allow us to bind a given alias to a given storage source we need
to save the index in virStorageSource. The backing store ids are now
generated when detecting the backing chain.

Since we don't re-detect the backing chain after snapshots, the
numbering needs to be fixed there.
Message-Id: <e25e817655585f926ecdef43aeb66280ad0109fd.1507835192.git.pkrempa@redhat.com>
Reviewed-by: Eric Blake <eblake@redhat.com>