Terminate backing chains explicitly
Express a properly terminated backing chain by putting a
virStorageSource of type VIR_STORAGE_TYPE_NONE in the chain. The newly
used helpers simplify this greatly.

The change fixes a bug as formatting an incomplete backing chain and
parsing it back would end up in expressing a terminated chain since
src->backingStoreRaw was not populated. By relying on the terminator
object this can be now processed appropriately.
Message-Id: <f864e5a1062e3ed550e7ada54e62ed8afb1e927c.1507835192.git.pkrempa@redhat.com>
Reviewed-by: Eric Blake <eblake@redhat.com>