conf: merge virDomainLifecycleCrashAction with virDomainLifecycleAction
There is no need to have two different enums where one has the same
values as the other one with some additions.

Currently for on_poweroff and on_reboot we allow only subset of actions
that are allowed for on_crash.  This was covered in parse time using
two different enums.  Now to make sure that we don't allow setting
actions that are not supported we need to check it while validating
domain config.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <be4cf1711e764c66d7039ae5f9021c4438ca3267.1508151956.git.phrdina@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>