qemu: implement virDomainSetLifecycleAction() API
There is one limitation for using this API, when the guest is started
with all actions set to "destroy" we put "-no-reboot" on the QEMU
command line.  That cannot be changed while QEMU is running and
the QEMU process is always terminated no matter what is configured
for any action.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1460677

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <1726b6aa3153d6915a2031c797ed5397bfda3476.1508151956.git.phrdina@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>