nodedev: udev: Introduce udevEventMonitorSanityCheck helper function
We need to perform a sanity check on the udev monitor before every
use so that we know nothing has changed in the meantime. The reason for
moving the code to a separate helper is to enhance readability and shift
the focus on the important stuff within the udevEventHandleCallback
handler.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <fe1f2137b48cc5ec60dd3b9da53bfb323daf11de.1508333637.git.eskultet@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>