util: Introduce virFileWaitForExists
Since we have a number of places where we workaround timing issues with
devices, attributes (files in general) not being available at the time
of processing them by calling usleep in a loop for a fixed number of
tries, we could as well have a utility function that would do that.
Therefore we won't have to duplicate this ugly workaround even more.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <aaf82946583079d777173506fb5ebe63350d9696.1508333637.git.eskultet@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>