qemu: remove pointless address validation on hot unplug
Back in the times of using 'pci_del', unplugging a device without
a PCI address was not wired up.

After completely removing support for qemu without QEMU_CAPS_DEVICE,
aliases are used to uniquely identify devices in all cases.

Remove the pointless validation of data that was already present
in the domain definition.
Message-Id: <c930ed855d4ae9b195bfe59e34de58920c326ccb.1508334928.git.jtomko@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>