qemu: domain: Simplify using DAC permissions of top of backing chain
qemuDomainGetImageIds and qemuDomainStorageFileInit are helpful when
trying to access a virStorageSource from the qemu driver since they
figure out the correct uid and gid for the image.

When accessing members of a backing chain the permissions for the top
level would be used. To allow using specific permissions per backing
chain level but still allow inheritance from the parent of the chain we
need to add a new parameter to the image ID APIs.
Message-Id: <24edeed8bcc9af56c9155a8823088204e9947681.1508507145.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>