qemu: domain: Extract setup for disk source secrets
Separate it so that it deals only with single virStorageSource, so that
it can later be reused for full backing chain support.

Two aliases are passed since authentication is more relevant to the
'storage backend' whereas encryption is more relevant to the protocol
layer. When using node names, the aliases will be different.
Message-Id: <593a3c3d8eee018e3172af5d7bbc42f68fd0df6d.1508507145.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>