qemu.conf: Introduce memory_predictable_file_names
In some cases management application needs to allocate memory for
qemu upfront and then just let qemu use that. Since we don't want
to expose path for memory-backend-file anywhere in the domain
XML, we can have a configuration knob that when enabled generated
predictable paths. In this case:

  $memoryBackingDir/libvirt/qemu/$shortName/$alias

where $shortName is result of virDomainObjGetShortName().

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <86a6f5b416296cf971e95ad09217da67fc5a7753.1508773227.git.mprivozn@redhat.com>