virsh: domain: Fix option handling in domxml-to-native
Commit fdeac7a05fdf85458d72e89efcfa0f444525aaad tried to fix the output
of 'virsh --help domxml-to-native' by switching types around. One of the
changes broke the option parser. VSH_OT_ARGV should be used only for
variable argument count, not to make the help generator look pretty.

The correct option type in this case is VSH_OT_STRING as it's not
mandatory now since it can be substituted by using --domain.

This makes --help for this command look incorrect, but the parser works
as it should.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1494400
Message-Id: <d5492596b0545e679150dd0f1f0812e530aeffd9.1508831491.git.pkrempa@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>