qemu: monitor: check monitor not closed upon send
Close monitor sets monitor error if another thread is awating the
response to propagate error condition to that thread. However
if there is no such thread error will not be set. Now if API
thread try to send a message it will hang. This can easily happen
for example if API thread does not reach the point when it take
domain lock and qemu driver is shutdowned.

Let's add checks for whether monitor is closed to send routine
and remove passing of this condition thru setting monitor error.
Message-Id: <1508922308-554701-5-git-send-email-nshirokovskiy@virtuozzo.com>