qemu: Use active and inactive snapshot configuration on restore
By default, active and inactive XMl snapshot configurations are
assigned to domain definition. This will make sure that all the
non-persistent configurations of the snapshot are restored back
as it is. This patch will also make sure that user has a choice
to choose of using active XML configuration of snapshot as both
active and inactive XML configurations of the restoring domain.

Signed-off-by: Kothapally Madhu Pavan <kmp@linux.vnet.ibm.com>
Message-Id: <1509353516-52208-3-git-send-email-kmp@linux.vnet.ibm.com>