RFC: qemu: add vmcoreinfo support
Starting from qemu 2.11, the `-device vmcoreinfo` will create a fw_cfg
entry for a guest to store dump details, necessary to process kernel
dump with KASLR enabled and providing additional kernel details.

Since the device is a singleton and shouldn't use additional hardware
resources, I decided to map this to a <feature> element in the libvirt
domain XML. Feel free to argue and decide if it is better to have it
under <devices> instead.

The device is arm/x86 only for now (targets that support fw_cfg+dma).

Related to:
https://bugzilla.redhat.com/show_bug.cgi?id=1395248

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>
Message-Id: <20171106120909.20382-1-marcandre.lureau@redhat.com>