vshCommandOpt: Relax check for valid options
When trying to get an opt for command typed on the command line
we first check if command has such option. Because if it doesn't
it is a programming error. For instance: vshCommandOptBool(cmd,
"config") called from say cmdStart() doesn't make sense since
there's no --config for the start command. However, we will want
to have generic completers which are going to check if various
options are set. And so it can happen that we will check for
non-existent option for given command. Therefore, we need to
relax the check otherwise we will hit the assert() and don't get
anywhere.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <9b9d2014b3ab57056532e46d8d1f0357956b8498.1510055948.git.mprivozn@redhat.com>