libvirtd: Alter refcnt processing for domain server objects
Whether the @srv/@srvAdm is added to the dmn->servers list or not,
the reference kept for the allocation can be dropped leaving just the
reference for being on the dmn->servers list be the sole deciding
factor when to really free the associated memory. The @dmn dispose
function (virNetDaemonDispose) will handle the Unref of the objects
during the virHashFree.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171108023956.8920-4-jferlan@redhat.com>