libvirtd: Alter refcnt processing for server program objects
Instead of holding onto the reference for each program added to
a NetServer, let's Unref the program reference as soon as the
attempt is made to add to the @srv->programs list. This then
allows the virNetServerDispose to handle the final Unref of
the object for actual deletion and cleans up the cleanup: path.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20171108023956.8920-5-jferlan@redhat.com>