conf: Properly parse <backingStore/>
The terminator would not be parsed properly since the XPath selector was
looking for an populated element, and also the code did not bother
assigning the terminating virStorageSourcePtr to the backingStore
property of the parent.

Some tests would catch it if there wasn't bigger fallout from the change
to backing store termination in a693fdba0111. Fix them properly now.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1509110
Message-Id: <35f2b726d96661e4a207218d2fc20c09eca2e131.1510152416.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>